Remove old interface and use Artifacts MUMPS_jll 5.4.1
