+++ Project.toml
@@ -7,11 +7,12 @@ version = "1.2.0"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
-SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
+MUMPS_jll = "ca64183c-ec4f-5579-95d5-17e128c21291"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
 MPI = "^0.7,^0.8,^0.9"
+MUMPS_jll = "= 5.4.1"
 julia = "^1.6"
 
 [extras]
